solutions/ydol/requirements.txt
requests
pycryptodome
beautifulsoup4
pypdf

